Selectable thumbnail dropdown
As a room owner, I would like to select thumbnails (out of all possible ones) so that thumbnails represent better the presentation.
In addition to the changes that I applied to _pubic_recording_row.html.erb, I implemented a dropdown list with thumbnails and captions (only first 20 letters). 
The room owner can select a thumbnail by clicking one of the png pics shown in the dropdown list. 
So far, one can select a thumbnail out of three possible candidates, a restriction of BBB which registers only the first three slides in the metadata.xml. I will post another PR in BBB which enables the registration of all possible thumbnail candidates.