Fix broken UI on long filenames
If you select a presentation with a very long filename in in Greenlight,
the filename text in the upload dialog wraps at the end and leaves the
label box, leading to a slightly broken looking user interface.

This patch adjusts the style to hide the end of a long filename, ending
with a horizontal ellipsis instead.