Add Sign In
Frontend:
User enter email and password in Sign In form.
If email and password correspond to a user in the server database, user is redirected to Home page on Form submit.
The information of the current user is displayed (name, email, provider, signed_in?).

Backend:
Server checks if User exists.
Server checks if User and Password match via authenticate method.
If success, return success status, signed_in as true and store User id in session.
If error, return error status and signed_in as false.